+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 ## [Unreleased]
+- `Langchain::LLM::Ollama` can now `#summarize`
+
+## [0.9.4]
+- New `Ollama#summarize()` method
+- Improved README
+- Fixes + specs
+
+## [0.9.3]
+- Add EML processor
+- Tools can support multiple-methods
+- Bump gems and bug fixes
 
 ## [0.9.2]
 - Fix vectorsearch#ask methods